Adding municipalities of Mexico
The official institution that regulates the codes and names is INEGI
http://www.inegi.org.mx/

INEGI defined two leading zeroes in the codes even when the municipality
count in some states doesn't exceed 99, for a total of three characters
in each code.

The state Chiapas doesn't use anymore the municipality code 095, because
it was merged in 1973 with another municipality.