Fix the data update script
The location and format of the ISO data has changed. The data is
available in plain JSON.

The fundamental data structure that is used by Carmen to build the YAML
files is not changed by this commit. This only tweaks the intermediary
step of getting the information from the Debian repository into the data
structure that Carmen needs.

ISO country data:
https://anonscm.debian.org/git/pkg-isocodes/iso-codes.git/plain/data/iso_3166-1.json

ISO region data:
https://anonscm.debian.org/git/pkg-isocodes/iso-codes.git/plain/data/iso_3166-2.json