Delete old YAML files before data update
This was not happening previously, so any regions/data files that were
out of date and no longer relevant were still living in the YAML
directories. Since we are pulling the entire ISO 3166-1 and ISO 3166-2
lists from the Debian repository we can safely build the entire data
directories from scratch each time.